Move data transformations to a separate file
